pyproject.toml
[tool.poetry]
name = "pushboards"
version = "0.1.0"
description = ""
authors = ["Viacheslav Matrizaev <matrizaev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dynaconf = "^3.2.4"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-login = "^0.6.3"
authlib = "^1.2.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
bootstrap-flask = "^2.3.3"
flask-wtf = "^1.2.1"
openpyxl = "^3.1.2"
pandas = "^2.1.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pylint = "^3.0.2"
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
pyproject-flake8 = "^6.1.0"
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120

[tool.pylint.format]
max-line-length = 120

[tool.flake8]
max-line-length = 120
exclude = ["migrations/*"]

[tool.pylint.main]
py-version = [ 3, 12 ]
disable = "C0114, C0115, C0116, R0903"
